move VideoElement to its own file
